(spike; DO NOT MERGE) add messages to expanded description
this doesn't produce anything in the spec tests other than a bunch of
newlines from all the other tests that didn't clear out their
descriptions

add_env isn't called until the block is called, which happens when the
task is called.

we might have to separate out our methods from the block.  but this
feels nasty, dunno if possible in a sustainable manner.